move PythonClientUser, rename StartBenchmarkResult to BenchmarkContext
